added creator field to update and delete posts

+++ backend/routes/posts.js
@@ -28,12 +28,17 @@ const storage = multer.diskStorage({
 });
 
 // post data in mongoDb
-router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
+router.post(
+  "",
+  checkAuth,
+  multer({storage: storage}).single("image"),
+  (req, res, next) => {
   const url = req.protocol + '://' + req.get("host");
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPosts => {
     res.status(201).json({
@@ -87,9 +92,12 @@ router.get("/:id", (req, res, next) =>{
 
 // delete a post
 router.delete("/:id", checkAuth, (req, res, next) => {
- Post.deleteOne({_id: req.params.id}).then(result => {
-   console.log(result);
-   res.status(200).json({message: "Post deleted successfully!"});
+ Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+  if(result.n > 0) {
+    res.status(200).json({message: "Post delete successfully!"});
+  } else {
+    res.status(401).json({message: "Unautherized!"});
+  }
  });
 });
 
@@ -107,8 +115,12 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"),
       content: req.body.content,
       imagePath: imagePath
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: "Post update successfully!"});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if(result.nModified > 0) {
+      res.status(200).json({message: "Post update successfully!"});
+    } else {
+      res.status(401).json({message: "Unautherized!"});
+    }
   });
 });
